==> 2-goodreads.ipynb <==
# Part 2: Goodreads Dataset

datasource: https://sites.google.com/eng.ucsd.edu/ucsdbookgraph/home

The data in the goodreads dataset is separated into four files: books, works, genres, and authors. We'll join these datasets and create a new dataset with relevant information for our analysis.

Since the files are large, we'll convert them to parquet format for faster processing.

import json
import polars as pl
import pandas as pd
from rapidfuzz import process
import pickle

# read in json files
genres = []
with open('./goodreads/json/goodreads genres.json') as f:
    for line in f:
        genres.append(json.loads(line))

genres_df = pd.DataFrame(genres)
genres_df['genres'] = genres_df['genres'].apply(lambda x: x.keys())
genres_df['genres'] = genres_df['genres'].apply(lambda x: [i for i in x if i != ''])
genres_df.to_parquet('./goodreads/genres.parquet')

authors_df = pl.read_ndjson('./goodreads/json/goodreads authors.json')

authorskeep = ['author_id', 'name']
authors_df = authors_df[authorskeep]

authors_df.write_parquet('./goodreads/authors.parquet')

author_id = authors_df['author_id'].to_list()
author_name = authors_df['name'].to_list()

author_dict = dict(zip(author_id, author_name))

# pickle author dict
with open('./goodreads/author_dict.pickle', 'wb') as handle:
    pickle.dump(author_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)

books_df = pl.read_ndjson('./goodreads/json/Goodreads Books.json')
books_df.write_parquet('./goodreads/books.parquet')

books_df = pl.read_parquet('./goodreads/books.parquet')

genres_df = pl.read_parquet('./goodreads/genres.parquet')

keep = [
    'country_code',
    'language_code',
    'average_rating',
    'description',
    'authors',
    'publisher',
    'num_pages',
    'publication_year',
    'image_url',
    'ratings_count',
    'book_id',
    'work_id',
    'title'
]

books_df = books_df[keep]

# join genres df to books df
books_df = books_df.join(genres_df, on='book_id', how='inner')

books_df.write_parquet('./goodreads/booksgenres.parquet')

books_df = pd.read_parquet('./goodreads/booksgenres.parquet')
authors_list = books_df['authors'].to_list()
with open('./goodreads/author_dict.pickle', 'rb') as handle:
    author_dict = pickle.load(handle)

for i in range(len(authors_list)):
    authors = []
    for j in authors_list[i]:
        authors.append(j['author_id'])
        authors_list[i] = authors

for i in range(len(authors_list)):
    for j in range(len(authors_list[i])):
        authors_list[i][j] = author_dict[authors_list[i][j]]

books_df['authors'] = authors_list

books_df['author_string'] = books_df['authors'].apply(lambda x: ', '.join(j for j in x))
books_df['author_title'] = books_df['author_string'] + ' - ' + books_df['title']

books_df.to_parquet('./goodreads/booksgenresauthors.parquet')

books_df.head()

books_df = pd.read_parquet('./goodreads/booksgenresauthors.parquet')

authortitle = books_df['author_title'].to_list()

# pickle author title list
with open('goodreads_titles.pickle', 'wb') as handle:
    pickle.dump(authortitle, handle, protocol=pickle.HIGHEST_PROTOCOL)
