==> 3-titlematching.ipynb <==
# Step 3: Matching titles

We'll use the rapidfuzz library to match titles from the Books3 dataset to the Goodreads dataset. We'll create a function that leverages Levenshtein distance to extract the best match for each title in the Books3 dataset.

from rapidfuzz import process
import polars as pl
import pandas as pd
import pickle
import tqdm

# read in books3 data and create list of titles
books_df = pl.read_parquet('books3copyright.parquet')
bookstitles = books_df['title'].to_list()

# read in goodreads titles pickle file
with open('./goodreads/authortitle.pickle', 'rb') as f:
    goodreads_titles = pickle.load(f)

# define matching function
def match_titles(title, titles_list):
    """
    Function to match titles from goodreads to books3 data
    """
    # use rapidfuzz to match titles
    match = process.extractOne(title, titles_list)
    # return the match and the score
    return title, match[0], match[1]
