==> 4-filtertables.ipynb <==
import pandas as pd
import polars as pl
import json


matches = pl.read_ndjson('matches.json').to_pandas()
books_df = pd.read_parquet('./goodreads/booksgenresauthors.parquet')

# if matches['match'] in books_df['author_title'], add to books_df row to new dataframe

matches_list = matches['match'].tolist()

books_df['match'] = books_df['author_title'].isin(matches_list)

books3_df = books_df[books_df['match'] == True]

books_df.to_csv('books3.tsv', sep='\t')

# drop duplicate author_title rows
books3_df = books3_df.drop_duplicates(subset=['author_title'], keep='first')

books3_df.to_json('books3.json', orient='records', lines=True)
