==> 5-findauthors.ipynb <==
import requests
from requests.adapters import HTTPAdapter, Retry
import pandas as pd
import pprint as pp
from bs4 import BeautifulSoup as bs
import spacy
import json
import time

nlp = spacy.load('en_core_web_trf')


books3_df = pd.read_json('books3.json', orient='records', lines=True)
books3_df.head()

authors = books3_df['author_string'].to_list()

headers = {
    'User-Agent': 'Research on Datasets (esuhr@gradcenter.cuny.edu)',
}

def make_request(url):
    session = requests.Session()
    retry = Retry(connect=3, backoff_factor=0.5)
    adapter = HTTPAdapter(max_retries=retry)
    session.mount('http://', adapter)
    session.mount('https://', adapter)
    response = session.get(url, headers=headers)
    parsed = response.json()
    return parsed

def findpages(author_name):
    author_name = author_name.replace(' ', '%20')
    url = f"https://en.wikipedia.org/w/api.php?action=query&format=json&list=search&srsearch={author_name}"
    data = make_request(url)
    if len(data['query']['search']) > 0:
        title = data['query']['search'][0]['title']
        return title

def gethtml(title):
    url = f"http://en.wikipedia.org/w/api.php?action=query&prop=revisions&rvprop=content&format=json&titles={title}&rvsection=0&rvparse"
    data = make_request(url)
    page_one = next(iter(data['query']['pages'].values()))
    if 'revisions' in page_one:
        html = next(iter(page_one['revisions'][0].values()))
        return html
    else:
        return None

def getdatelocs(html):
    if html is None:
        return None, None
    else:
        soup = bs(html, 'html.parser')
        # find span class="bday"
        infobox = soup.find('td', {'class': 'infobox-data'})
        if infobox:
            info = ''.join(infobox.text)
            info = info.replace('(', ' ').replace(')', ' ')

            infolabels = nlp(info)
            if 'DATE' in [ent.label_ for ent in infolabels.ents] and 'GPE' in [ent.label_ for ent in infolabels.ents]:
                date = [ent.text for ent in infolabels.ents if ent.label_ == 'DATE'][0]
                location = [ent.text for ent in infolabels.ents if ent.label_ == 'GPE']
                locationstr = ', '.join([str(loc) for loc in location])
                return date, locationstr
            else:
                return None, None
        else:
            return None, None

authorinfo = []
for i in range(len(authors)):
    authordict = {}
    author = authors[i]
    time.sleep(1)
    title = findpages(author)
    time.sleep(1)
    html = gethtml(title)
    date, location = getdatelocs(html)
    print(f'{i}:{author} was born on {date} in {location}')
    authordict['author'] = author
    authordict['bday'] = date
    authordict['location'] = location
    with open('authorinfo.json', 'a') as f:
        json.dump(authordict, f)
        f.write('\n')
